requirements.txt
openai>=1.0.0
selenium>=4.0.0
webdriver-manager>=3.8.0
python-dotenv>=1.0.0
beautifulsoup4==4.13.3
requests
selenium-stealth